Homepage visits no longer cause writes to runs collection
Runs finish when the last task finishes
- Not when it's viewed after the last task finishes
- Fixed an auto-purge bug that creates more tasks than expected
- Changed auto-purge to run for tests using more than 1 thread
- stop_run is called after all tasks are finished
  - Run auto-purge before marking the run as finished
  - If no workers were purged, mark the run as finished
  - Calculate and save results_info when the run is finished
  - Set the run status fields (finished, is_green, etc.)

Moved misc helper functions to util.py
- Any shared helper function that doesn't persist to the db
- Move format_results, bad worker detection, etc.
- Move estimate_game_duration

Moved some views.py functions to rundb methods
- Move purge_run method
- Move auto-purge code
- Move aggregate_unfinished_runs
- Enable filtering by username in get_unfinished_runs

Updates to tests
- Add missing num_iter field to test
- Add more tests and comments
- Test that results_info is set correctly
- Use chunk size in tests instead of hardcoded 250
